refactor(bindings): extract buildCommandsArray to sibling .mjs
Matches the repo convention for pure-JS runtime helpers (resolveRegion.mjs,
matchRoute.mjs, lruCache.mjs) and lets the rebind test import the helper
directly instead of doing `new Function( source-slice )` extraction from
the BindingsConsumer.js source string.

- New `src/runtime/bindings/buildCommandsArray.mjs` owns the helper.
  Imports parseShortcut from its sibling. Pure ESM.
- `BindingsConsumer.js` imports + re-exports the helper (re-export keeps
  the public surface stable for any existing consumers).
- `tests/runtime/bindings-consumer-rebind.test.mjs` imports the helper
  directly. Header comment now honestly flags that the four static-
  analysis assertions guard the *shape* of the perf fix, not its
  *behavior* — a future legitimate `compiled` → `entries` rename would
  trip them (accepted trade-off). Also lists the three coverage gaps
  that need JSDOM mount tests (deferred to issue #30).

The `helper extracts from source` assertion is gone (no longer needed
since the test imports the helper directly). Net delta: -1 assertion on
this test.
